Prevent running makepkg for AUR packages on subsequent preparation runs
A previous preparation run might have populated the source directory with a
PKGBUILD file but not an .SRCINFO file, e.g. due to an error when writing
the latter. So far, the build preparation would have invoked makepkg for
the PKGBUILD from the AUR on a subsequent run in the absence of an
.SRCINFO file. With this change this is prevented by alway writing a
"from-aur" file to the source directory (before writing any other files)
and not invoking makepkg if this file is present.

This way the source directory can still be populated with invalid or
partial data but this is actually useful for investigation in such an error
case.